+++ src/tests/components/resetPassword.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { shallow } from "enzyme";
-import { resetpassword as PasswordReset, mapStateToProps } from "../../components/resetpassword";
+import { resetpassword as PasswordReset, mapStateToProps } from "../../components/resetpassword1";
 let resetPasswordComp;
 const token =
   "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImRhbm55a2Ftb3NvMjAyM0BnbWFpbC5jb20iLCJpYXQiOjE1OTE2NTcxNzEsImV4cCI6MTU5MTY2MDc3MX0.aUwB5Q6TF6KjM8y6viuhQlnAbW_z3oji2v9g689dp1c";
